not sure what I did in src/.  tired!
anyway, coded a basic C template system for the web.  it still
needs a few more features (like an include directive), as well
as code for an apache module.

I also need to code more of the runtime, especially support
for accessing a database (mongodb?).

the parser is overkill; the template file format could be parsed
by a few regexp's.  it does, however, mean I can extent the grammar
in literally any way I want in the future, though.